+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-4271] Permissions for linking implementations with business concepts
+
 ## [4.40.0] 2022-03-14
 
 ### Changed
@@ -7,6 +13,7 @@
 - [TD-2501] Database timeout and pool size can now be configured using
   `DB_TIMEOUT_MILLIS` and `DB_POOL_SIZE` environment variables
 - [TD-4491] Refactored search and permissions
+
 
 ## [4.28.0] 2021-09-20